scale pieChart accurately and most other graphs
breaks when the div that it's getting scaled to is greater than the original image size. Seems to be relative to how large the layout is (which is customizable)

currently scales by only one gradient.. either the scaling difference between the length OR the width. Using both causes the graph to distort lengthwise or widthwise. This is because of the set aspect ratio in dashboardView. (could probably be changed/fixed but things render very well currently)